Don't overwrite previously recorded tests results if another publishing was called but no files were found by the provided filter
There is a bug - if you call nunit results publisher twice and if
first publish was successful and recorded some results and the
second publish is executed for nonexisting file - the result will be
overriden and will be "No tests results". Don't call recordTestResult in
case if there are no nunit tests results to don't overwrite existing
test results. This is specific when "failIfNoResults" is false.