Fix parallel publishing of NUnit test results
Before this change if you try to publish test results in parallel it
would probably make something wrong - add two test result actions,
overwrite the existing test results, messing junit temporary reports.
Changed this to be able to process reports in parallel. Applied the
similar change as was applied in the JUnit plugin.
https://github.com/jenkinsci/junit-plugin/commit/3fd230f488f85a1094f34a5e5143c69c00ff59ce