Add the failedTestsMakeBuildUnstable pipeline parameter
With this we'll be able to tell nunit plugin to don't make build
unstable in case of tests failure. You may want to set this if you
publish some tests results with failed tests and you want this build to
be successful, in this case you need to don't make this build "UNSTABLE"
because you can't change build result from "worse" to "better" in
Jenkins.